applied height config for Grading page, implemented common AppPage class for loading config
