+++ pages/Config.py
@@ -4,66 +4,22 @@ import shutil
 import streamlit as st
 import toml
 
+from pages.Page import AppPage
 
-class ConfigPage:
+
+class ConfigPage(AppPage):
     def __init__(self):
-        self.CONFIG_PATH = os.path.join(os.path.dirname(os.path.dirname(__file__)), ".streamlit", "config.toml")
-        self.default_config = {
-            "save": {"dir": "/Users/naoki/OneDrive - Kyoto University/assignments"},
-            "window": {"grading_height": 740},
-        }
-        self.config = self.load_config()
-
+        super().__init__()
         st.session_state.setdefault("just_saved", False)
         if st.session_state.get("just_saved"):
             st.toast("設定を保存しました。", icon="✅")
             st.session_state["just_saved"] = False
-
-    def load_config(self):
-        if not os.path.exists(self.CONFIG_PATH):
-            os.makedirs(os.path.dirname(self.CONFIG_PATH), exist_ok=True)
-            with open(self.CONFIG_PATH, "w") as f:
-                toml.dump(self.default_config, f)
-            return self.default_config
-        try:
-            curr_config = toml.load(self.CONFIG_PATH)
-            return self.merge_dicts(self.default_config, curr_config)
-        except Exception:
-            return self.default_config
 
     def save_config(self):
         """Saves the current configuration to the config file."""
         with open(self.CONFIG_PATH, "w") as f:
             toml.dump(self.config, f)
         print("Saved current config", self.config)
-
-    def merge_dicts(self, default, override):
-        """
-        Recursively merges two dictionaries, with values from the override dictionary
-        taking precedence. If both dictionaries contain a value for the same key and
-        both values are dictionaries, they are merged recursively.
-
-        Parameters
-        ----------
-        default : dict
-            The base dictionary to merge into.
-        override : dict
-            The dictionary whose values will override those in the base dictionary.
-
-        Returns
-        -------
-        dict
-            A new dictionary containing the merged keys and values.
-        """
-        result = default.copy()
-        for key, value in override.items():
-            if key in result and isinstance(result[key], dict) and isinstance(value, dict):
-                # ネストされた辞書を再帰的にマージ
-                result[key] = self.merge_dicts(result[key], value)
-            else:
-                # 単純な値や型が異なる場合は上書き
-                result[key] = value
-        return result
 
     def has_assignments(self, path):
         if not os.path.isdir(path):
@@ -149,13 +105,11 @@ class ConfigPage:
             step=10,
             key="window_height",
         )
-        self.config["window"]["grading_height"] = height
-        st.button(
-            "保存",
-            on_click=self.save_config,
-            key="save_height_btn",
-        )
-        st.session_state["just_saved"] = True
+        if st.button("保存", key="save_height_btn"):
+            # Update the config with the new height
+            self.config["window"]["grading_height"] = height
+            self.save_config()
+            st.session_state["just_saved"] = True
 
     def render(self):
         st.title("Configuration")
